+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -6,8 +6,7 @@ import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.coyote.http11.controller.AbstractController;
-import org.apache.coyote.http11.controller.Controller;
-import org.apache.coyote.http11.cookie.HttpCookie;
+import org.apache.coyote.http11.common.HttpCookie;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.RequestBody;
 import org.apache.coyote.http11.response.HttpResponse;
