+++ tomcat/src/main/java/org/apache/coyote/http11/common/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.cookie;
+package org.apache.coyote.http11.common;
 
 import java.util.ArrayList;
 import java.util.HashMap;
