+++ tomcat/src/test/java/org/apache/coyote/http11/cookie/HttpCookieTest.java
@@ -1,5 +1,6 @@
 package org.apache.coyote.http11.cookie;
 
+import org.apache.coyote.http11.common.HttpCookie;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
